+++ frontend/src/helpers/Globals.tsx
